requirements.txt
absl-py==1.4.0
appdirs==1.4.4
asttokens==2.2.1
audioread==3.0.0
av==10.0.0
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cmake==3.26.0
decorator==4.4.2
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.9.0
fsspec==2023.4.0
future==0.18.3
fvcore @ git+https://github.com/facebookresearch/fvcore@51092b5515cbb493f73de079743dd6b11cc4bbf1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
huggingface-hub==0.13.4
idna==3.4
imageio==2.28.0
imageio-ffmpeg==0.4.8
iopath==0.1.10
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
lazy_loader==0.2
librosa==0.10.0.post2
lit==16.0.0
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.2.1
msgpack==1.0.5
mypy-extensions==1.0.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai-whisper @ git+https://github.com/openai/whisper.git@c09a7ae299c4c34c5839a76380ae407e7d785914
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parameterized==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pooch==1.6.0
portalocker==2.7.0
proglog==0.1.10
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyre-extensions==0.0.30
python-dateutil==2.8.2
pytorchvideo==0.1.5
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.3.1
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.1+cu118
torchdim @ git+https://github.com/facebookresearch/torchdim@b2a058ab5ada860f63fe48eadc98e7373dbdd9ec
torcheval==0.0.6
torchtnt==0.0.7
torchvision==0.15.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
triton==2.0.0
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.2.3
yacs==0.1.8
